update schema config
